Got code ready for full integration video

+++ src/pico_FreeRTOS/src/queues.h
@@ -5,7 +5,7 @@ struct message
 {
     double latitude;
     double longitude;
-    double speed;
+    int speed;
     int message_type;
 };
